Corrigindo especialidade, agora sim completo

+++ disque-saude-backend/src/main/java/com/ufcg/si1/repository/EspecialidadeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ufcg.si1.model.Especialidade;
 
 @Repository
-public interface EspecialidadeRepository extends JpaRepository<Especialidade, Integer> {
+public interface EspecialidadeRepository extends JpaRepository<Especialidade, Long> {
 
 }
